Change the type of SparseMatrix::num_nonzeroes and SparseMatrix::max_num_nonzeroes from int to int64_t.
This will allow for larger size Jacobians to be allocated, when
memory is available and matrix dimensions are still 32bit.

+++ internal/ceres/dense_sparse_matrix.cc
@@ -67,7 +67,7 @@ DenseSparseMatrix::DenseSparseMatrix(const TripletSparseMatrix& m)
   const double *values = m.values();
   const int *rows = m.rows();
   const int *cols = m.cols();
-  int num_nonzeros = m.num_nonzeros();
+  int64_t num_nonzeros = m.num_nonzeros();
 
   for (int i = 0; i < num_nonzeros; ++i) {
     m_(rows[i], cols[i]) += values[i];
@@ -137,7 +137,7 @@ int DenseSparseMatrix::num_cols() const {
   return m_.cols();
 }
 
-int DenseSparseMatrix::num_nonzeros() const {
+int64_t DenseSparseMatrix::num_nonzeros() const {
   if (has_diagonal_reserved_ && !has_diagonal_appended_) {
     return (m_.rows() - m_.cols()) * m_.cols();
   }
